+++ modules/acm/outputs.tf
@@ -15,7 +15,6 @@
  */
 
 output "git_creds_public" {
-  description = "Public key of SSH keypair to allow the Anthos Operator to authenticate to your Git repository."
-  value       = var.create_ssh_key ? tls_private_key.git_creds.*.public_key_openssh : null
+  description = "Public key of SSH keypair to allow the Anthos Config Management Operator to authenticate to your Git repository."
+  value       = module.acm_operator.git_creds_public
 }
-