+++ modules/private-cluster/versions.tf
@@ -18,6 +18,6 @@ terraform {
   required_version = "~> 0.12.6"
 
   required_providers {
-    google = ">= 2.18, <4.0.0"
+    google = ">= 3.16, <4.0.0"
   }
 }