+++ test/integration/beta_cluster/controls/gcloud.rb
@@ -60,6 +60,9 @@ control "gcloud" do
           "cloudRunConfig" => {},
           "dnsCacheConfig" => {
             "enabled" => true,
+          },
+          "gcePersistentDiskCsiDriverConfig" => {
+            "enabled" => true,
           }
         })
       end
@@ -71,8 +74,9 @@ control "gcloud" do
       end
 
       it "has the expected nodeMetadata conseal config" do
-        expect(data['nodeConfig']['workloadMetadataConfig']).to include({
-          "nodeMetadata" => 'EXPOSE',
+        expect(data['nodeConfig']['workloadMetadataConfig']).to eq({
+          "mode" => "GKE_METADATA",
+          "nodeMetadata" => 'GKE_METADATA_SERVER',
         })
       end
 
@@ -86,6 +90,13 @@ control "gcloud" do
         expect(data['databaseEncryption']).to eq({
           "state" => 'ENCRYPTED',
           "keyName" => attribute('database_encryption_key_name'),
+        })
+      end
+
+      it "has the expected workload identity config" do
+        expect(data['workloadIdentityConfig']).to eq({
+          "identityNamespace" => attribute('identity_namespace'),
+          "workloadPool" => attribute('identity_namespace'),
         })
       end
     end
@@ -196,6 +207,19 @@ control "gcloud" do
           including(
             "management" => including(
               "autoRepair" => true,
+            ),
+          )
+        )
+      end
+
+      it "has the expected node metadata for workload identity" do
+        expect(node_pools).to include(
+          including(
+            "config" => including(
+              "workloadMetadataConfig" => eq(
+                "mode" => "GKE_METADATA",
+                "nodeMetadata" => 'GKE_METADATA_SERVER',
+              ),
             ),
           )
         )